backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Supabase & HTTP ────────────────────────────────────────────
supabase>=2.4.0
httpx>=0.27.0

# ── Auth & config ──────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.7.0

# ── ML core ────────────────────────────────────────────────────
# IMPORTANT: pinned to 1.7.2 — model_v2.pkl was pickled with this version.
# Upgrading sklearn without retraining the model causes InconsistentVersionWarning
# and silent prediction failures (everything becomes "Other").
scikit-learn==1.7.2
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0
joblib>=1.4.0
openpyxl>=3.1.0

# ── Transformer embeddings ─────────────────────────────────────
sentence-transformers>=3.0.0

# ── Gmail fetch ───────────────────────────────────────────────
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.128.0

# ── Async task queue (Removed) ──────────────────────────────────────────

# ── CRON Jobs ─────────────────────────────────────────────────
apscheduler
pytz